Demonstrate workaround for SPR-8582
The structure of this project as originally submitted introduces a
fundamental chicken-and-egg project that cannot be 'fixed' per se.

Config tries to use @Autowired to inject TheBeanWithMessageSource,
which itself is defined in XML.  This is not usually a problem, but
TheBeanWithMessageSource depends on a bean (messageSource) that is
defined within Config.  Because @Autowired processing must occur upon
instantiation of the Config class (before the messageSource @Bean method
can be processed), it is impossible to fulfill this particular request:
the MessageSource that TheBeanWithMessageSource needs is currently in
creation and thus the 'circular reference' error is raised.

XML-based configuration does not incur this problem, because it is not
constrained by the rules of the Java language and runtime.  In this way
XML remains more flexible, but of course it does not have the advantages
that Java offers.  With that said, it should be mentioned that the
scenario demonstrated in this issue has rarely come up, so should
probably not be considered a major limitation.

Kai's original workaround of simply creating a second @Configuration
class to get around the circular reference is reasonable. This commit
submits another way to get there: instead of using @Autowired at the
field level, use parameter injection at the @Bean method level (see
the commit diff to see what I mean).  This gets around all lifecycle
issues, and may be just what you need.

Another way to make this happen would be to @Inject or @Autowire the
ApplicationContext itself (or implement ApplicationContextAware) and
then look up TheBeanWithMessageSource when it is needed (within the
otherBean() method).