Simplify and demonstrate workaround for SPR-8639
As originally submitted, SPR-8639 used ProxyFactoryBean to apply
MyInterceptor to beans A, B and C.

Typically this works fine, but because of a subtle circular dependency
being introduced due to A and B depending on C and C needing to autowire
D, the PFBs for A and B are prematurely queried for their type (to see
if they are autowire candidates for C) and this trigger a
BeanAlreadyInCreation exception that is subsequently suppressed.

This is a very specific use case and highly unlikely to be reproduced
by other users with any frequency.  However, we will address it even if
only by allowing the original exception to propagate fully.

In the meantime, I have refactored the test case to demonstrate use of
BeanNameAutoProxyCreator in lieu of the original ProxyFactoryBean, and
this approach is better in every way: it is more concise, more readable,
requires fewer bean definitions, and best of all does not suffer from
the above-mentioned bug.

Issue: SPR-8639