Introduce unwrapAopProxyIfNecessary util method
For reasons mentioned in the previous commit, detecting/unwrapping AOP
proxies within AbstractAutowireCapableBeanFactory is not feasible. The
work here rather needs to be done by the client.

The idea is simply to use AopUtils to detect and unwrap the target of
an AOP proxy *before* calling beanFactory.autowireBeanProperties(...),
and this commit demonstrates how to do so.

Note that use of #autowireBeanProperties, #configureBean and other
similar methods are primarily designed for use by third-party
integrators as opposed to everyday Spring developers. For this reason
it's justifiable to require a little more effort to get things working.