Add repro project for SPR-9176
Project demonstrates that use of @Resource for fields within a
request-scoped component results in a memory leak, because the
injected object (in this case an HttpServletRequest) is proxied, and
the bean name is not tracked by the dependent bean map because the name
is different for each injection, e.g. $Proxy10@12345 vs $Proxy10@98765.

The workaround for this issue is to use @Autowired instead of @Resource,
but naturally the @Resource leak needs to be addressed.