Update and refresh template projects
This commit reviews the infrastructure provided by the the war
template projects. The web application can now be easily deployed in
a servlet container with support for the most obvious ones (jetty
and tomcat).

Maven warnings have been fixed and the overall structure of the
build script has been harmonized

The READMEs have been migrated to markdown and the new infrastructure
has been documented.