Test case for SPR-11723
The test is actually for a variant of SPR-11430 raised as an issue
in Spring Boot: https://github.com/spring-projects/spring-boot/issues/710.

In this form there is an @ImportResource() where the XML contains a
@Configuration that is not processed. It seems even more
surprising to users than the vanilla SPR-11430 problem because
(as the test shows) the @Configuration *is* processed if it
is not imported via @ImportResource (e.g. if the XML is
loaded directly).