Add sample code for child contexts in test methods
I blended my original ideas with Thomas's comments and got something
that works with a throwaway context per method if it detects either
@ChildSpringApplication or @ChildTestProperties (separate annotations
needed because the existing analogs at the type level do not
support being added to methods).

There's some reflection hackery to get the TestContext in a MethodRule,
but possibly that would be unnecessary if this feature was supported
in SpringMethodRule.